tests: reverse some test skip logic
Some of the skip tests check for if the test has to be skipped first,
i.e, the test is skipped in the if statement (pass), and the test is
tested in the else statement (fail).

This reverses some skip test logics so all the tests look similar and
expect to be skipped in the if statement (and thus, to be tested in the
else statement).