tests: reverse some test logic
Most of the tests check for the success conditions, i.e, the test
succeeds in the if statement (pass), and the test fails in the else
statement (fail):

	if cqfd ...; then # or ! cqfd ...
		echo "yeah, test has passed!"
	else
		echo "oops, test has FAILED!"
	fi

However, some tests do not:

	if cqfd ...; then # or ! cqfd ...
		echo "oops, test has FAILED!"
	else
		echo "yeah, test has passed!"
	fi

This turns the test logic to "if all needed conditions to succeed the
test are met, then the test passes, otherwise if any of the conditions
is unmet, then the test fails".

It reverses some test logics so all the tests look similar and expect to
pass in the if statement (and thus, fail in the else statement).