tests: 04-cqfd_init_context: simplify tests
The two tests testing if the build_context is set or unset grep for the
build_cont in the container file /tmp/cqfdrc-build_context.

That file is added if by the Dockerfile.build_context[1] is used (it is
not if the Dockerfile is used). The .cqfdrc file cqfdrc-build_context[2]
sets build_context=".".

The first test uses the standard Dockerfile and it does not set the
build_context= (the context is directory holding the Dockerfile, i.e.
the project directory .cqfd/docker). The file /tmp/cqfdrc-build_context
is not added in the container.

	debug executing: docker build -t cqfd_gportay_cqfd_test_XXXXXXX /tmp/tmp.XXXXXXXXXX/.cqfd/docker

The second test uses the alternate Dockerfile.build_context and it sets
the build_context= to the "current" working directory (i.e. the project
top directory). The file /tmp/cqfdrc-build_context is added in the
container.

	debug executing: docker build -t cqfd_gportay_cqfd_context_test_XXXXXXX . -f /tmp/tmp.XXXXXXXXXX/.cqfd/docker/Dockerfile

Note: The current working directory . is /tmp/tmp.XXXXXXXXXX/

This simplifies the two tests by testing the presence of the file
/tmp/cqfdrc-build_context in the container instead of grepping a pattern
in the file that is not necessary present.

[1]: https://github.com/savoirfairelinux/cqfd/blob/9299823ce661fefa1ac17d6501fde408cedd6a14/tests/test_data/.cqfd/docker/Dockerfile.build_context#L7
[2]: https://github.com/savoirfairelinux/cqfd/blob/9299823ce661fefa1ac17d6501fde408cedd6a14/tests/test_data/cqfdrc-build_context#L4