tests: 04-cqfd_init_context: fix testing if cqfd init fail
The first test uses the standard Dockerfile and it does not set the
build_context= (the context is directory holding the Dockerfile, i.e.
the project directory .cqfd/docker).

The file /tmp/cqfdrc-build_context is not added in the container.

The test is set to fail if cqfd init succeeds **AND** if the file exists
in the container (cqfd run).

Thus, the test passes if cqfd init fails (or if the file is not in the
container), but, the test **MUST** fails if cqfd init fails.

This fixes the test by turning the logic to "if all needed conditions to
succeed the test are met, then the test passes, otherwise if any of the
conditions is unmet, then the test fails".