cqfd: run -c, exec: check arguments after shifting
The run command option -c and the command exec both requires at least
one argument.

cqfd checks for the remaining number of arguments, including the run
option -c or the exec command, and thus, it checks against 2, while it
appears more simple to check against l (>= 1) or 0 (> 0).

This checks the number of arguments after shifting the run command
option -c or the command exec; it is easier to read how much arguments
these commands require

Note: This output the usage if arguments are missing.