cqfd: rename function config_load()
cqfd names the most of the functions in the format <verb>[_<noun>]():
 - make_archive()
 - make_launcher()
 - locate_project_dir()

The function config_load() is in the reverse format.

This renames the function config_load() to load_config() to

Note: The docker functions are named after their underlying docker
command (docker_build(), docker_run()). The usage(), warn(), die(),
debug() functions are somehow "standard". The function cfg_parser() is
taken and modified from[1].

[1]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/