tests: 05-cqfd_init_custom_image: reuse TDIR
The tests are run in a mktemp directory already[1][2], and that
directory is removed at the end of tests[3] (if the tests are not
interrupted). The directory is pointed by "$TDIR".

This reuses the TDIR to remove the use of a duplicated mktemp to create
an alternate temporary cqfd structure. It backs up the cqfd files and
restores them at the end of test.

This avoid poluting the temporary directory of the system if the tests
are interrupted.

[1]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L14
[2]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L17
[3]: https://github.com/savoirfairelinux/cqfd/blob/v5.7.0/tests/Makefile#L29-L31