tests: 05-cqfd_init_custom_image: test pulling inexistent image fails
cqfd with custom_img_name pulls the image if it does not exist in the
local registry; bypassing cqfd init.

Note: cqfd init builds the custom_img_name using the Dockerfile in the
.cqfd directory.

This adds a test to test the failure of pulling of an inexistant image
from distant registries; the custom image is previously removed.