cqfd: allow relative and absolute path with -d option
The option -f allows absolute .cqfdrc while the option -d allows
directory name "only".

The directory name given to the option -d is used by the function
locate_project_dir() to locate the project directory (i.e. the .cqfd
parent directory). The project directory searches for the directory name
backward starting by the current working directory up to root.

Important: At that point, relative directory works, but it is not
suitable to use more path components than a (single) filename. Using
cqfd -d subdir/to/dot-cqfd works fine but the function searches for the
directory "subdir/to/dot-cqfd" and this is not suitable if using option
-C to change directory; cqfd -C subdir/to -d dot-cqfd is preferable.

This allows giving relative and absolute path to option -d to specify
the path to the .cqfd directory to use, by hacking the function
locate_project_dir() to return the parent directory of the .cqfd set by
the option -d, and by prepending the working directory to the cqfddir
variable if the option -d is unset.

Fixes: #173