cqfd: add cqfd start command
This adds 'cqfd start' command. Which is to start a container created
using the 'no-rm' option.

Note that the command should have container id as an argument.

To achieve this, when container run with the option, the entrypoint is
created inside ~/.cache/cqfd/containers.

This also adds the cqfd_user_home and cqfd_user_cwd declaration inside
the load_config function to avoid duplication.