cqfd: add cqfd store command
This implements the 'cqfd store' command. This is to store a running
container to an image. This takes two arguments, ID of the running
container and the image name where the container to be stored.

This command is doing a 'docker commit container_id image_name'. This
will help developers to have a snapshot of a cqfd container at a given
state. Which is helpful for testing purpose.