tests: move to bats
This commit makes all the tests compatible with the bats framework. This
ensures a more standard way to test cqfd, compared to the homemade jtest
framework.

Bats tests often have the .bats extension, this is why we added the
`.bats` extension to all tests, as well as to differentiate them from
the old version