podman: better support
GitHub Actions do not expose the variable $GID so we make it default to
$UID instead

The --userns=keep-id option maps the files belonging to the user who is
launching the podman command to the same uid in the container. This is
what we want. However, this option creates an entry in the /etc/passwd
file for the current user, but setting its home directory to `/` and the
entry is not added if another user has the same UID, which is the case
for example with ubuntu 24.04 container where there is already an ubuntu
user with UID 1000. So we have two cases, if there is the ubuntu user
then no entry is added, and if UID is available, then an entry with the
wrong home path is added. This is why we first delete the line in
/etc/passwd which corresponds to the user UID and then we add the user
using `useradd` to have the options we want

The --user=root option allows to launch the container first as a root
user, which allows us for example to set the home directory of the
current user to match the one on the host. The entrypoint script takes
care of doing that