cqfd: use CQFD_HISTFILE
The variable HISTFILE is not exported in most shells so cqfd cannot read
it. And if we don’t use the same shell on the host and in cqfd then we
don’t want to share the histfile. So it is more convenient to use this
variable to set the HISTFILE in cqfd